requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.32.0
tortoise-orm[asyncpg]==0.21.6
aerich==0.7.2
celery==5.4.0
redis==5.2.0
aiohttp==3.10.10
jinja2==3.1.4
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1